\name{f1}
\alias{f1}
\title{Returns the input incremented by one}
\usage{f1(x, logger=create.logger(NULL))}
\description{Given any input, returns it plus 1}
\value{\code{x+1}}
\arguments{
  \item{x}{Anything goes, but if `+` is not implemented on it an error
    is returned}
  \item{logger}{An instance of class log4r:::logger}
}
