\name{f2}
\alias{f2}
\title{Returns the input decremented by one}
\usage{f2(x, logger=create.logger(NULL))}
\description{Given any input, returns it minus 1}
\value{\code{x-1}}
\arguments{
  \item{x}{Anything goes, but if `-` is not implemented on it an error
    is returned}
  \item{logger}{An instance of class log4r:::logger}
}
